trying to test the catheter table and sort out the circular imports that are a mess
