+++ brachyutils/geometry/__init__.py
@@ -24,7 +24,7 @@ from .egsphant_utils import _load_json
 from .applicator_utils import BrachyApplicator
 
 # trunk-ignore(ruff/F401)
-from .catheter_utils import DwellPosition, Catheter, CatheterTable
+# from .catheter_utils import DwellPosition, Catheter, CatheterTable
 
 # trunk-ignore(ruff/F401)
 from .registration_utils import BrachyPhantomRegistration
