+++ brachyutils/geometry/catheter_utils/catheter_table.py
@@ -1,8 +1,8 @@
 import copy
 import numpy as np
-from typing import List, Union, Dict, Any, Optional, Tuple
+from typing import List, Union, Dict, Any, Optional, Tuple, Literal
 from pathlib import Path
-from pydantic import BaseModel, computed_field, ConfigDict, model_validator
+from pydantic import BaseModel, computed_field, ConfigDict, model_validator, Field
 import json
 import SimpleITK as sitk
 from opentps.core.processing.imageProcessing.sitkImageProcessing import imageToSITK
@@ -21,7 +21,31 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 from tqdm import tqdm
 from collections import defaultdict
 from itertools import chain
-from brachyutils.planning.plan_utils import ExportConfig_Dose
+
+class ExportConfig_Dose(BaseModel):
+    """
+    Configuration for exporting dose data from the plan.
+    """
+    model_config = ConfigDict(
+        arbitrary_types_allowed=True,
+        use_attribute_docstrings=True  # Enables auto-docs from Field desc [web:48]
+    )
+    dir_export: str | Path = Field(None, description="Directory where dose files are exported.")
+    name_combined: str = Field("combined", description="File name for combined dose output.")
+    file_extension: Literal[".seq.nrrd", ".3ddose"] = Field(
+        ".seq.nrrd", description="Allowed file extensions for dose files."
+    )
+    write_dose_rate_maps: bool = Field(
+        False, description="Whether to write individual dose rate maps to files."
+    )
+    multi_processing: bool = Field(
+        True, description="Enable multiprocessing for export (yes/no toggle)."
+    )
+    @computed_field
+    def pth_combined(self)->Path:
+        self.dir_export = Path(self.dir_export)
+        return self.dir_export/(self.name_combined+self.file_extension)
+
 
 class DwellPosition(BaseModel):
     r"""
